Use executable-relative paths when importing datasets in demo
Instead of relying on the current worker directory which may not
always work (e.g., when demo is run from the MSVS), assume that
the datasets folder is in the same folder as the demo executable.

In other words, do the same as is done when loading SVGs for
bitmap bundles.