Rewrite FileDownload to not store request and response objects
When a request fails, it seems to go into a fugue state as its destroyed by the even handler (even though we have a copy of it). Now, just use these on demand and also cache response info as separate values.

Also, since FileDownload is synchronous, remove the CancelPending function.